Portable hashing
We can lean on the defined return type of SXHASH for the definition
of the type. Mixing is less trivial, and this code will only work
very well for 64 bit arithmetic. It should still work if an
implementation's fixnums are bigger or smaller, though with less
mixing or with bignum consing respectively.